feat (Peyvandtel Service Prices): added `index` and `store` request
the index API will show a paginated list of service prices.
the store API will validate data using each service validator if any, and it will create a new serve price for that service. for now the relation between service and service provider is one to one but it can be change to one to many if needed.
the ServicePriceTest file test various scenarios of creating a service price.
each service validator will throw the specified ServiceException to be able to call `apiError` macro of response.