feat (Part AI): send files shorter than 60 seconds with another API (performance)
sending files shorter than 60 seconds with `file` API instead of `largeFile`. The new API will instantly send the text of that voice in the response. added new attributes to the model, serviceDTO and  other classes in order to detect the final result is available or not, if so, send the text in the `STORE` API of services.